Pathfinding logic finished
The pure logic of pathfinding works, now we need to improve the quality of the pathing and add weights to simulate a good AI pathfinder. (e.g. if there is mud on the ground, the unit will try to avoid it, as long as it is worth to do it)